pyproject.toml
[project]
name = "elk-logger"
version = "0.1.0"
description = "Logger for ELK with beautiful stdout output"
requires-python = ">=3.11"
dependencies = [
    "python-logstash-async>=2.5.0",
    "colorlog>=6.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"